refactor: refar Page component form components
