+++ src/YupTrialNested.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import { useFormik } from 'formik';
 import * as Yup from 'yup';
-import Page from './Page';
+import Page from './components/Page';
 import { Typography } from '@mui/material';
 interface FormikValuesProps {
   name: string;
